pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian-timemachine"
version = "0.1.0"
description = "Automated backup solution for Obsidian Vault with Git version control"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Jacky" }
]
keywords = ["obsidian", "backup", "git", "rsync", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Archiving :: Backup",
]

dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
ot = "ot.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["ot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
